% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablePrepare.R
\name{close.prepared}
\alias{close.prepared}
\title{Closes an object created with tablePrepare}
\usage{
\method{close}{prepared}(con, ...)
}
\arguments{
\item{con}{object of class \code{prepared}, i.e. created with \code{\link{tablePrepare}}.}

\item{...}{not used}
}
\value{
\code{TRUE} if the files could be closed, \code{FALSE} if it was closed already
}
\description{
Closes an object created with tablePrepare. It frees the Memory Mapping resources and closes the underlying \code{ff} data without deleting the file data.
}
