% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_loadPrepare.R
\name{loadPrepare}
\alias{loadPrepare}
\title{Loads a prepared object}
\usage{
loadPrepare(dir)
}
\arguments{
\item{dir}{directory of the prepared object}
}
\value{
the prepared object
}
\description{
Loads a prepared object that has been saved with \code{\link{savePrepare}}.
}

