% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tabplot.R
\name{print.tabplot}
\alias{print.tabplot}
\title{Print a \link{tabplot-object}}
\usage{
\method{print}{tabplot}(x, ...)
}
\arguments{
\item{x}{tabplot object}

\item{...}{arguments passed to other methods}
}
\description{
Print a \link{tabplot-object}
}

