\name{summary.tabplot}
\alias{summary.tabplot}
\title{Summary of a tabplot-object.}
\usage{summary.tabplot(object, digits=max(3, getOption("digits") - 3), ...)
}
\description{Summary of a \link{tabplot-object}.}
\alias{summary.tabplot}
\arguments{\item{object}{tabplot object}
\item{digits}{integer, used for number formatting with \code{\link{format}()}}
\item{...}{arguments passed to other methods}
}

