\name{tableGUI_showAllPals}
\alias{tableGUI_showAllPals}
\title{Show all palettes of the tabplot package...}
\description{Show all palettes of the tabplot package}
\details{All palettes in \code{\link{tabplotPalettes}} are shown in a new device.}
\alias{tableGUI_showAllPals}
\keyword{color}
\keyword{palettes}

