\name{tableplot_checkBins}
\alias{tableplot_checkBins}
\title{Function to check the tableplot argument: number of bins.}
\usage{
  tableplot_checkBins(nBins, nDat)
}
\arguments{
  \item{nBins}{the number of bins}

  \item{nDat}{the number of data observations}
}
\value{
  the (possibly corrected) number of bins
}
\description{
  Function to check the tableplot argument: number of bins.
}

