\name{tableplot_checkCols}
\alias{tableplot_checkCols}
\title{Function to check the tableplot argument: sortCol}
\usage{
  tableplot_checkCols(sortCol, colNames)
}
\arguments{
  \item{sortCol}{sortCol}

  \item{colNames}{colNames}
}
\value{
  (possibly corrected) sortCol
}
\description{
  Function to check the tableplot argument: sortCol
}

