\name{tableplot_checkDecreasing}
\alias{tableplot_checkDecreasing}
\title{Function to check the tableplot argument: decreasing}
\usage{
  tableplot_checkDecreasing(decreasing, sortCol)
}
\arguments{
  \item{decreasing}{decreasing}

  \item{sortCol}{sortCol}
}
\value{
  (possibly corrected) value of decreasing
}
\description{
  Function to check the tableplot argument: decreasing
}

