\name{tableplot_checkFromTo}
\alias{tableplot_checkFromTo}
\title{Function to check the tableplot arguments: from and to.}
\usage{
  tableplot_checkFromTo(from, to)
}
\arguments{
  \item{from}{from}

  \item{to}{to}
}
\description{
  Function to check the tableplot arguments: from and to.
}

