\name{tableplot_checkPals}
\alias{tableplot_checkPals}
\title{Function to check the tableplot argument: pals}
\usage{
  tableplot_checkPals(pals)
}
\arguments{
  \item{pals}{pals}
}
\value{
  object with palette information
}
\description{
  Function to check the tableplot argument: pals
}

