\name{tableplot_showPalettes}
\alias{tableplot_showPalettes}
\title{Show all palettes of the tabplot package...}
\description{Show all palettes of the tabplot package}
\details{All palettes in \code{\link{tabplotPalettes}} are shown in a new device.}
\alias{tableplot_showPalettes}
\keyword{color}
\keyword{palettes}

