\name{tabplot_compare-object}
\alias{tabplot_compare-object}
\title{object that contains the information to plot the difference of two tableplots (experimental)}
\description{
A \code{tabplot_compare} is created by substracting two
tableplots (see \link{tableplot_comparison}). For numeric
variables, both absolute and relative difference of the
mean values are computed. For categorical variables, the
freqency tables are compared.
}

