\name{tabplotPalettes}
\alias{tabplotPalettes}
\alias{tabplotPalettes}
\title{Qualitative palettes used for tableplot...}
\description{Qualitative palettes used for tableplot}
\details{Color brewer palettes:
\itemize{
\item default 
\item Accent
\item Dark2
\item Paired
\item Pastel1
\item Pastel2
\item Set1
\item Set2
\item Set3
}

HCL palettes:
\itemize{
\item hcl5
\item hcl8
\item hcl12
}

Color blind friendly palette:
\itemize{
\item col_blind_friendly
}}
\alias{tabplotPalettes}
\docType{data}
\references{\url{http://colorbrewer2.org/}
\url{http://jfly.iam.u-tokyo.ac.jp/color/}}

