\name{datetime2fac}
\alias{datetime2fac}
\title{Function that creates a factor vector out of a date-time vector.}
\usage{datetime2fac(p)
}
\description{Function that creates a factor vector out of a date-time vector.}
\details{Creates a factor vector out of a date-time vector that an inherits from class "POSIXt" or "Date".}
\alias{datetime2fac}
\value{A factor vector}
\note{\code{ff} vectors are not implemented yet}
\arguments{\item{p}{date-time vector}
}

