\name{plot.tabplot}
\alias{plot.tabplot}
\title{Function that plots a tabplot-object.}
\usage{plot.tabplot(x, fontsize=8, legend.lines=8, ...)
}
\description{Function that plots a \link{tabplot-object}.}
\alias{plot.tabplot}
\arguments{\item{x}{tabplot object}
\item{fontsize}{the (maximum) fontsize}
\item{legend.lines}{the number of lines preserved for the legend}
\item{...}{arguments passed to other methods}
}

