\name{summary.tabplot}
\alias{summary.tabplot}
\title{Summarize a \link{tabplot-object}}
\usage{
\method{summary}{tabplot}(object, digits = max(3, getOption("digits") - 3),
  ...)
}
\arguments{
  \item{object}{tabplot object}

  \item{digits}{integer, used for number formatting with
  \code{\link{format}}}

  \item{...}{arguments passed to other methods}
}
\description{
Summarize a \link{tabplot-object}
}

